// SOKO-main/src/types.ts
export interface PostVideoData {
  videoUrl?: string;
  thumbnailUrl: string;
  duration: string;
  durationSeconds?: number;
  viewsCount: string;
  speakerName?: string;
  speakerTitle?: string;
  speakerAvatar?: string;
  badge?: string;
  subtitles?: { time: number; text: string }[];
  chapters?: { timeSeconds: number; timestamp: string; title: string }[];
}